Initial commit
Code style is all over the place because I wrote this quickly
and didn't set up R# yet.

Run `OrbitalCommand.exe mytagname`  from inside a git repo to
get the short message for all commits going back to `mytagname`,
with relevant bug numbers next to them. For example:

    Fix crash (12345)
    Fix style nit
    Fix two related issues (23456, 34567)

Bug numbers are found based commit messages lines starting with
"Fixes:" and followed by a FogBugz URL.

The idea is to output some pretty markdown that can be quickly
edited and used in the body of a Github Release, for example.
Not quite there yet.